+++ test/request_queue.js
@@ -278,7 +278,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', undefined, clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             const requestA = new Request({ url: 'http://example.com/a' });
@@ -288,6 +289,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestA,
                     forefront: false,
+                    clientId,
                 })
                 .returns(Promise.resolve({ requestId: 'a', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestA }));
             await queue.addRequest(requestA);
@@ -299,6 +301,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestB,
                     forefront: true,
+                    clientId,
                 })
                 .returns(Promise.resolve({ requestId: 'b', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestB }));
             await queue.addRequest(requestB, { forefront: true });
@@ -336,6 +339,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestB,
                     forefront: true,
+                    clientId,
                 })
                 .returns(Promise.resolve({ requestId: requestB.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestB }));
             await queue.reclaimRequest(requestB, { forefront: true });
@@ -362,6 +366,7 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     request: requestB,
+                    clientId,
                 })
                 .returns(Promise.resolve({ requestId: requestB.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestB }));
             await queue.markRequestHandled(requestB);
@@ -374,6 +379,7 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     items: [
@@ -411,7 +417,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', undefined, clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             const requestA = new Request({ url: 'http://example.com/a' });
@@ -424,6 +431,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestA,
                     forefront: false,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     requestId: 'a',
@@ -451,7 +459,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', undefined, clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             const requestX = new Request({ url: 'http://example.com/x' });
@@ -464,6 +473,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestX,
                     forefront: false,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     requestId: 'x',
@@ -491,7 +501,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id', 'some-name');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', 'some-name', clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             // Query queue head with request A
@@ -500,6 +511,7 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     items: [
@@ -527,7 +539,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id', 'some-name');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', 'some-name', clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
             mock.expects('getHead').never();
 
@@ -540,6 +553,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestA,
                     forefront: true,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     requestId: 'a',
@@ -578,7 +592,8 @@ describe('RequestQueue', () => {
 
             const { Request } = Apify;
 
-            const queue = new RequestQueue('some-id', 'some-name');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', 'some-name', clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             const requestA = new Request({ url: 'http://example.com/a' });
@@ -589,6 +604,7 @@ describe('RequestQueue', () => {
                     queueId: 'some-id',
                     request: requestA,
                     forefront: true,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     requestId: 'a',
@@ -603,6 +619,7 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
                 })
                 .returns(Promise.resolve({ items: [] }));
 
@@ -615,7 +632,8 @@ describe('RequestQueue', () => {
 
         it('should accept plain object in addRequest()', async () => {
             expectNotUsingLocalStorage();
-            const queue = new RequestQueue('some-id');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', undefined, clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
             mock.expects('addRequest')
                 .once()
@@ -623,6 +641,7 @@ describe('RequestQueue', () => {
                     requestId: 'xxx',
                     wasAlreadyHandled: false,
                     wasAlreadyPresent: false,
+                    clientId,
                 }));
             await queue.addRequest({ url: 'http://example.com/a' });
             mock.verify();
@@ -642,10 +661,11 @@ describe('RequestQueue', () => {
             sinon.restore();
         });
 
-        it('should always wait for a queue head to become consistent before marking queue as finished', async () => {
+        it('should always wait for a queue head to become consistent before marking queue as finished (hadMultipleClients = true)', async () => {
             expectNotUsingLocalStorage();
 
-            const queue = new RequestQueue('some-id', 'some-name');
+            const clientId = 'my-client';
+            const queue = new RequestQueue('some-id', 'some-name', clientId);
             const mock = sinon.mock(apifyClient.requestQueues);
 
             // Return head with modifiedAt = now so it will retry the call.
@@ -654,11 +674,13 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     limit: 5,
-                    queueModifiedAt: new Date(),
+                    queueModifiedAt: new Date(Date.now() - API_PROCESSED_REQUESTS_DELAY_MILLIS * 0.75),
                     items: [],
+                    hadMultipleClients: true,
                 }));
 
             // And now return return date which makes the queue consistent.
@@ -667,11 +689,134 @@ describe('RequestQueue', () => {
                 .withArgs({
                     queueId: 'some-id',
                     limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
                 })
                 .returns(Promise.resolve({
                     limit: 5,
                     queueModifiedAt: new Date(Date.now() - API_PROCESSED_REQUESTS_DELAY_MILLIS),
                     items: [],
+                    hadMultipleClients: true,
+                }));
+
+            expect(await queue.isFinished()).to.be.eql(true);
+
+            mock.verify();
+            mock.restore();
+        });
+
+        it('should always wait for a queue head to become consistent before marking queue as finished (hadMultipleClients = true)', async () => {
+            expectNotUsingLocalStorage();
+
+            const clientId = 'my-client';
+            const queueId = 'some-id';
+            const queue = new RequestQueue(queueId, 'some-name', clientId);
+            const mock = sinon.mock(apifyClient.requestQueues);
+
+            expect(queue.assumedTotalCount).to.be.eql(0);
+            expect(queue.assumedHandledCount).to.be.eql(0);
+
+            // Add some requests.
+            const requestA = new Apify.Request({ url: 'http://example.com/a' });
+            const requestB = new Apify.Request({ url: 'http://example.com/b' });
+            mock.expects('addRequest').once()
+                .withArgs({ queueId, request: requestA, forefront: true, clientId })
+                .returns(Promise.resolve({ requestId: 'a', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestA }));
+            mock.expects('addRequest').once()
+                .withArgs({ queueId, request: requestB, forefront: true, clientId })
+                .returns(Promise.resolve({ requestId: 'b', wasAlreadyHandled: false, wasAlreadyPresent: false, request: requestA }));
+            await queue.addRequest(requestA, { forefront: true });
+            await queue.addRequest(requestB, { forefront: true });
+
+            expect(queue.queueHeadDict.length()).to.be.eql(2);
+            expect(queue.inProgressCount).to.be.eql(0);
+            expect(queue.assumedTotalCount).to.be.eql(2);
+            expect(queue.assumedHandledCount).to.be.eql(0);
+
+            // It won't query the head as there is something in progress or pending.
+            mock.expects('getHead').never();
+            expect(await queue.isFinished()).to.be.eql(false);
+
+            // Fetch them from queue.
+            mock.expects('getRequest').once()
+                .withArgs({ queueId: 'some-id', requestId: 'b' })
+                .returns(Promise.resolve(_.extend(requestB, { id: 'b' })));
+            mock.expects('getRequest').once()
+                .withArgs({ queueId: 'some-id', requestId: 'a' })
+                .returns(Promise.resolve(_.extend(requestA, { id: 'a' })));
+            const requestBFromQueue = await queue.fetchNextRequest();
+            expect(requestBFromQueue).to.be.eql(requestB);
+            const requestAFromQueue = await queue.fetchNextRequest();
+            expect(requestAFromQueue).to.be.eql(requestA);
+
+            expect(queue.queueHeadDict.length()).to.be.eql(0);
+            expect(queue.inProgressCount).to.be.eql(2);
+            expect(queue.assumedTotalCount).to.be.eql(2);
+            expect(queue.assumedHandledCount).to.be.eql(0);
+
+            // It won't query the head as there is something in progress or pending.
+            mock.expects('getHead').never();
+            expect(await queue.isFinished()).to.be.eql(false);
+
+            // Reclaim one and mark another one handled.
+            mock.expects('updateRequest').once()
+                .withArgs({ queueId, request: requestB, clientId })
+                .returns(Promise.resolve({ requestId: requestB.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestB }));
+            await queue.markRequestHandled(requestB);
+            mock.expects('updateRequest').once()
+                .withArgs({ queueId, request: requestA, forefront: true, clientId })
+                .returns(Promise.resolve({ requestId: requestA.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestA }));
+            await queue.reclaimRequest(requestA, { forefront: true });
+
+            expect(queue.queueHeadDict.length()).to.be.eql(1);
+            expect(queue.inProgressCount).to.be.eql(0);
+            expect(queue.assumedTotalCount).to.be.eql(2);
+            expect(queue.assumedHandledCount).to.be.eql(1);
+
+            // It won't query the head as there is something in progress or pending.
+            mock.expects('getHead').never();
+            expect(await queue.isFinished()).to.be.eql(false);
+
+            // Fetch again.
+            mock.expects('getRequest').once()
+                .withArgs({ queueId: 'some-id', requestId: 'a' })
+                .returns(Promise.resolve(_.extend(requestA, { id: 'a' })));
+            const requestAFromQueue2 = await queue.fetchNextRequest();
+            expect(requestAFromQueue2).to.be.eql(requestA);
+
+            expect(queue.queueHeadDict.length()).to.be.eql(0);
+            expect(queue.inProgressCount).to.be.eql(1);
+            expect(queue.assumedTotalCount).to.be.eql(2);
+            expect(queue.assumedHandledCount).to.be.eql(1);
+
+            // It won't query the head as there is something in progress or pending.
+            mock.expects('getHead').never();
+            expect(await queue.isFinished()).to.be.eql(false);
+
+            // Mark handled.
+            mock.expects('updateRequest').once()
+                .withArgs({ queueId, request: requestA, clientId })
+                .returns(Promise.resolve({ requestId: requestA.id, wasAlreadyHandled: false, wasAlreadyPresent: true, request: requestA }));
+            await queue.markRequestHandled(requestA);
+
+            expect(queue.queueHeadDict.length()).to.be.eql(0);
+            expect(queue.inProgressCount).to.be.eql(0);
+            expect(queue.assumedTotalCount).to.be.eql(2);
+            expect(queue.assumedHandledCount).to.be.eql(2);
+
+            // Return head with modifiedAt = now so it would retry the query for queue to become consistent but because hadMultipleClients=true
+            // it will finish immediately.
+            mock.expects('getHead')
+                .once()
+                .withArgs({
+                    queueId: 'some-id',
+                    limit: QUERY_HEAD_MIN_LENGTH,
+                    clientId,
+                })
+                .returns(Promise.resolve({
+                    limit: 5,
+                    queueModifiedAt: new Date(),
+                    items: [],
+                    hadMultipleClients: false,
                 }));
 
             expect(await queue.isFinished()).to.be.eql(true);