+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxxxxxxxxxxxxxxxxxx
+===================
+- Request queue now allows crawlers to finish quickly without waiting in a case that queue was used by a single client.
+
 0.13.5 / 2019-03-27
 ===================
 - Fix `RequestQueue` throttling at high concurrency.
@@ -28,7 +32,7 @@
 - **BREAKING CHANGE:** Added `handleRequestTimeoutSecs` option to `BasicCrawler` with a 60 second default.
 - **DEPRECATED:** `PuppeteerPool` options in the `PuppeteerCrawler` constructor are now deprecated.
   Please use the new `puppeteerPoolOptions` argument of type `Object` to pass them. `launchPuppeteerFunction`
-  and `launchPuppeteerOptions` are still available as shortcuts for convenience.   
+  and `launchPuppeteerOptions` are still available as shortcuts for convenience.
 - `CheerioCrawler` and `PuppeteerCrawler` now automatically set `handleRequestTimeoutSecs` to 10 times
   their `handlePageTimeoutSecs`. This is a precaution that should keep requests from hanging forever.
 - Added `options.prepareRequestFunction()` to `CheerioCrawler` constructor to enable modification